+++ spec/dummy/app/controllers/inertia_test_controller.rb
@@ -27,6 +27,32 @@ class InertiaTestController < ApplicationController
     redirect_to :empty_test
   end
 
+  def external_redirect_test
+    redirect_to 'http://external-website.com/some_path', allow_other_host: true
+  end
+
+  def external_redirect_with_cookie_test
+    cookies['external_cookie'] = 'hello'
+    redirect_to 'http://external-website.com/some_path', allow_other_host: true
+  end
+
+  def same_origin_redirect_test
+    redirect_to empty_test_url
+  end
+
+  def secure_same_host_redirect_test
+    redirect_to "https://#{request.host}/empty_test", allow_other_host: true
+  end
+
+  def external_redirect_with_inertia_errors_test
+    redirect_to 'http://external-website.com/some_path', allow_other_host: true, inertia: { errors: { uh: 'oh' } }
+  end
+
+  def location_header_test
+    response.headers['Vary'] = params[:vary] if params[:vary].present?
+    head (params[:status] || 302).to_i, location: params[:url]
+  end
+
   def inertia_request_test
     if request.inertia?
       head 202
@@ -63,6 +89,10 @@ class InertiaTestController < ApplicationController
   # https://github.com/rails/rails/issues/28033
   def my_location
     inertia_location empty_test_path
+  end
+
+  def my_external_location
+    inertia_location 'http://external-website.com/some_path'
   end
 
   def redirect_with_inertia_errors